+++ src/main/java/com/jwplayer/jwplatform/client/OriginalClient.java
@@ -1,0 +1,5 @@
+package com.jwplayer.jwplatform.client;
+
+public class OriginalClient {
+
+}